Improve build speed by adding a base docker image
