+++ .github/dependabot.yml
@@ -6,7 +6,7 @@ updates:
     interval: daily
   open-pull-requests-limit: 10
 - package-ecosystem: pip
-  directory: "/binder"
+  directory: "/docker/binder"
   schedule:
     interval: daily
   open-pull-requests-limit: 10
