+++ README.md
@@ -18,12 +18,12 @@ Some versions have to be updated in multiple places:
 
 ### Nessie
 
-Nessie version is set in Binder at `binder/requirements.txt` and for unit tests in `notebooks/tox.ini`. Currently, Iceberg and Delta
+Nessie version is set in Binder at `docker/binder/requirements.txt` and for unit tests in `notebooks/tox.ini`. Currently, Iceberg and Delta
 both support only 0.9.x of Nessie.
 
 ### Iceberg
 
-Currently we are using Iceberg `0.12.0` and it is specified in both iceberg notebooks as well as `notebooks/tests/__init__.py`
+Currently we are using Iceberg `0.12.0` and it is specified in both iceberg notebooks as well as `docker/utils/__init__.py`
 
 ### Delta
 
@@ -31,22 +31,22 @@ currently Delta version is taken directly from the Nessie version and isn't expl
 
 ### Spark
 
-Only has to be updated in `binder/requirements.txt`. Currently Iceberg supports 3.0.x and 3.1.x while delta late supports
+Only has to be updated in `docker/binder/requirements.txt`. Currently Iceberg supports 3.0.x and 3.1.x while delta late supports
 3.1.x only.
 
 ### Flink
 
-Flink version is set in Binder at `binder/postBuild` and for unit tests in `notebooks/tox.ini`. Currently, Iceberg supports
+Flink version is set in Binder at `docker/binder/postBuild` and for unit tests in `notebooks/tox.ini`. Currently, Iceberg supports
 only 1.12.1
 
 ### Hadoop
 
-Hadoop libs are used by flink and currently specified in `notebooks/tests/__init__.py` only. We use 2.10.1 with Flink.
+Hadoop libs are used by flink and currently specified in `docker/utils/__init__.py` only. We use 2.10.1 with Flink.
 
 ### Hive
 
 Current Hive version that is being used `2.3.9` which supports Hadoop version of `2.10.1`. To update the version, it needs to be only updated 
-in `notebooks/tests/__init__.py`.
+in `docker/utils/__init__.py`.
 
 ## Binder
