+++ docker/binder/resources/hive/config/hive-site.xml
