+++ notebooks/tests/__init__.py
@@ -17,41 +17,10 @@
 #
 """Unit tests for demo notebooks."""
 import os
-import shutil
-import site
-import stat
 import subprocess  # noqa: S404
-import sysconfig
-import tarfile
 from contextlib import contextmanager
-from typing import Optional
 
-import requests
-
-try:
-    import pyspark
-
-    _SPARK_VERSION = pyspark.__version__
-    _SPARK_FILENAME = f"spark-{_SPARK_VERSION}-bin-hadoop3.2"
-    _SPARK_URL = f"https://archive.apache.org/dist/spark/spark-{_SPARK_VERSION}/{_SPARK_FILENAME}.tgz"
-except ImportError:
-    _SPARK_VERSION = None
-    _SPARK_FILENAME = None
-    _SPARK_URL = None
-
-_HADOOP_VERSION = "2.10.1"
-_HADOOP_FILENAME = f"hadoop-{_HADOOP_VERSION}"
-_HADOOP_URL = f"https://archive.apache.org/dist/hadoop/common/hadoop-{_HADOOP_VERSION}/{_HADOOP_FILENAME}.tar.gz"
-
-_ICEBERG_VERSION = "0.12.0"
-_ICEBERG_FLINK_FILENAME = f"iceberg-flink-runtime-{_ICEBERG_VERSION}.jar"
-_ICEBERG_FLINK_URL = f"https://repo1.maven.org/maven2/org/apache/iceberg/iceberg-flink-runtime/{_ICEBERG_VERSION}/{_ICEBERG_FLINK_FILENAME}"
-_ICEBERG_HIVE_FILENAME = f"iceberg-hive-runtime-{_ICEBERG_VERSION}.jar"
-_ICEBERG_HIVE_URL = f"https://repo1.maven.org/maven2/org/apache/iceberg/iceberg-hive-runtime/{_ICEBERG_VERSION}/{_ICEBERG_HIVE_FILENAME}"
-
-_HIVE_VERSION = "2.3.9"
-_HIVE_FILENAME = f"apache-hive-{_HIVE_VERSION}-bin"
-_HIVE_URL = f"https://apache.mirror.digionline.de/hive/hive-{_HIVE_VERSION}/{_HIVE_FILENAME}.tar.gz"
+from utils import fetch_nessie
 
 
 def _find_notebook(notebook_file: str) -> str:
@@ -65,133 +34,6 @@ def _find_notebook(notebook_file: str) -> str:
             f"Could not find {notebook_file} in {os.path.abspath('.')} and {os.path.abspath('..')}"
         )
     return os.path.abspath(path_to_notebook)
-
-
-def _get_unzip(filename: str, url: str) -> None:
-    if not os.path.exists(filename):
-        response = requests.get(url, stream=True)
-        with tarfile.open(fileobj=response.raw, mode="r|gz") as file:
-            file.extractall(path=".")
-
-
-def fetch_spark() -> None:
-    """Download and unzip Spark."""
-    if not _SPARK_VERSION:
-        raise NotImplementedError(
-            "Can't download spark as pyspark hasn't been installed."
-        )
-    _get_unzip(_SPARK_FILENAME, _SPARK_URL)
-    os.environ["SPARK_HOME"] = os.path.join(os.getcwd(), _SPARK_FILENAME)
-
-
-def _get_hadoop() -> None:
-    _get_unzip(_HADOOP_FILENAME, _HADOOP_URL)
-    os.environ["HADOOP_HOME"] = os.path.join(os.getcwd(), _HADOOP_FILENAME)
-
-
-def _copy_all_hadoop_jars_to_pyflink() -> None:
-    _get_hadoop()
-    if not os.getenv("HADOOP_HOME"):
-        raise Exception(
-            "The HADOOP_HOME env var must be set and point to a valid Hadoop installation"
-        )
-
-    pyflink_lib_dir = _find_pyflink_lib_dir()
-    duplicates = 0
-    for _jar_count, jar in enumerate(_jar_files()):
-        try:
-            shutil.copy(jar, pyflink_lib_dir)
-        except FileExistsError:
-            duplicates += 1
-            print(f"Duplicate jar {jar}")
-    print(
-        f"Copied {_jar_count} HADOOP jar files into the pyflink lib dir at location {pyflink_lib_dir} with {duplicates} duplicates"
-    )
-
-
-def _find_pyflink_lib_dir() -> Optional[str]:
-    for dir in site.getsitepackages() + [sysconfig.get_paths()["purelib"]]:
-        package_dir = os.path.join(dir, "pyflink", "lib")
-        if os.path.exists(package_dir):
-            return package_dir
-    return None
-
-
-def _jar_files() -> str:
-    for root, _, files in os.walk(os.getenv("HADOOP_HOME")):
-        for file in files:
-            if file.endswith(".jar"):
-                yield os.path.join(root, file)
-
-
-def _get(filename: str, url: str) -> None:
-    if os.path.exists(filename):
-        return filename
-    r = requests.get(url)
-    with open(filename, "wb") as f:
-        f.write(r.content)
-
-
-def fetch_nessie() -> str:
-    """Download nessie executable."""
-    runner = "nessie-quarkus-runner"
-    import pynessie
-
-    version = pynessie.__version__
-    url = "https://github.com/projectnessie/nessie/releases/download/nessie-{}/nessie-quarkus-{}-runner".format(
-        version, version
-    )
-    _get(runner, url)
-    os.chmod(runner, os.stat(runner).st_mode | stat.S_IXUSR)
-    return runner
-
-
-def fetch_iceberg_flink() -> str:
-    """Download flink jar for iceberg."""
-    filename = _ICEBERG_FLINK_FILENAME
-    url = _ICEBERG_FLINK_URL
-    _get(filename, url)
-    return filename
-
-
-def fetch_hive() -> None:
-    """Download and unzip Hive."""
-    _get_unzip(_HIVE_FILENAME, _HIVE_URL)
-    os.environ["HIVE_HOME"] = os.path.join(os.getcwd(), _HIVE_FILENAME)
-
-
-def fetch_iceberg_hive() -> str:
-    """Download Hive jar for iceberg."""
-    filename = _ICEBERG_HIVE_FILENAME
-    url = _ICEBERG_HIVE_URL
-    _get(filename, url)
-    return filename
-
-
-def fetch_hive_with_iceberg_jars() -> None:
-    """Download both Hive and Iceberg Hive jars."""
-    fetch_hive()
-    if not os.getenv("HIVE_HOME"):
-        raise Exception(
-            "The HIVE_HOME env var must be set and point to a valid Hive installation"
-        )
-
-    hive_auxlib_dir = os.path.join(os.getenv("HIVE_HOME"), "auxlib")
-
-    if not os.path.exists(hive_auxlib_dir):
-        os.mkdir(hive_auxlib_dir)
-
-    iceberg_hive_jar = fetch_iceberg_hive()
-
-    try:
-        shutil.copy(iceberg_hive_jar, hive_auxlib_dir)
-    except FileExistsError:
-        print(f"Jar {iceberg_hive_jar} exists already.")
-
-
-def fetch_hadoop() -> None:
-    """Download Hadoop jars."""
-    _get_hadoop()
 
 
 @contextmanager
