+++ notebooks/tests/scripts/start_hive
@@ -1,15 +1,15 @@
 #!/bin/bash
 
-python -c "import tests;tests.fetch_hadoop()"
+python -c "import utils;utils.fetch_hadoop()"
 
-python -c "import tests;tests.fetch_hive_with_iceberg_jars()"
+python -c "import utils;utils.fetch_hive_with_iceberg_jars()"
 
-HADOOP_VERSION=`python -c "import tests;print(tests._HADOOP_VERSION)"`
+HADOOP_VERSION=`python -c "import utils;print(utils._HADOOP_VERSION)"`
 
-HIVE_VERSION=`python -c "import tests;print(tests._HIVE_VERSION)"`
+HIVE_VERSION=`python -c "import utils;print(utils._HIVE_VERSION)"`
 
 export HADOOP_HOME=$PWD/hadoop-$HADOOP_VERSION
 
 #Start Hive
-chmod +x $PWD/../binder/start.hive
-nohup $PWD/../binder/start.hive $PWD $PWD/../binder/resources $HIVE_VERSION
+chmod +x $PWD/../docker/binder/start.hive
+nohup $PWD/../docker/binder/start.hive $PWD $PWD/../docker/binder/resources $HIVE_VERSION
