+++ notebooks/tox.ini
@@ -33,7 +33,7 @@ commands =
 
 [testenv:flink]
 setenv =
-    PYTHONPATH = {toxinidir}
+    PYTHONPATH = {toxinidir}:{toxinidir}/../docker
 passenv = TOXENV CI TRAVIS TRAVIS_* CODECOV_*
 deps =
     findspark
@@ -49,10 +49,10 @@ commands =
 whitelist_externals=
     /bin/bash
 setenv =
-    PYTHONPATH = {toxinidir}
+    PYTHONPATH = {toxinidir}:{toxinidir}/../docker
 passenv = TOXENV CI TRAVIS TRAVIS_* CODECOV_* JAVA_HOME
 deps =
-    -r{toxinidir}/../binder/requirements.txt
+    -r{toxinidir}/../docker/binder/requirements.txt
     -r{toxinidir}/requirements_dev.txt
     -r{toxinidir}/requirements.txt
 commands =
@@ -63,7 +63,7 @@ commands =
 
 [testenv]
 setenv =
-    PYTHONPATH = {toxinidir}
+    PYTHONPATH = {toxinidir}:{toxinidir}/../docker
 passenv = TOXENV CI TRAVIS TRAVIS_* CODECOV_*
 deps =
     -r{toxinidir}/../binder/requirements.txt
